fix token and network saving as well as circular dependency issues (#661)
