+++ apps/main-landing/src/app/creators/utils/navigation.ts
@@ -1,12 +1,36 @@
 'use client';
 
-import { useCallback } from 'react';
+import { SetStateAction, useCallback } from 'react';
 import { useRouter } from 'next/navigation';
-import { usePrivy } from '@privy-io/react-auth';
+import { getAccessToken, usePrivy, User } from '@privy-io/react-auth';
+import { Hex } from 'viem';
+import { CREATOR_API_URL } from '@idriss-xyz/constants';
 
 import { useAuth } from '../context/auth-context';
 
-import { setCreatorIfSessionPresent } from '.';
+import { CreatorProfileResponse } from './types';
+
+export const getCreatorProfile = async (
+  authToken: string,
+): Promise<CreatorProfileResponse | undefined> => {
+  if (!authToken) {
+    console.error('No slug to get creator');
+    return;
+  }
+
+  const response = await fetch(`${CREATOR_API_URL}/creator-profile/me`, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': `Bearer ${authToken}`,
+    },
+  });
+  if (!response.ok) {
+    return;
+  }
+  const data = (await response.json()) as CreatorProfileResponse;
+  return data;
+};
 
 export const useStartEarningNavigation = () => {
   const router = useRouter();
@@ -30,4 +54,23 @@ export const useStartEarningNavigation = () => {
   }, [user, router, creator, setIsModalOpen, setCreator]);
 
   return handleStartEarningClick;
+};
+
+export const setCreatorIfSessionPresent = async (
+  user: User,
+  setCreator: (value: SetStateAction<CreatorProfileResponse | null>) => void,
+) => {
+  const walletAddress = user.wallet?.address as Hex | undefined;
+  const authToken = await getAccessToken();
+  if (!authToken || !user.id) {
+    throw new Error('Could not get auth token or user ID for new user.');
+  }
+  const fetchedCreator = await getCreatorProfile(authToken);
+  if (fetchedCreator == null || fetchedCreator === undefined) {
+    console.error(
+      `Could not find creator with address ${walletAddress}, this may be a new user.`,
+    );
+    return;
+  }
+  setCreator(fetchedCreator);
 };
